cms/requirements.txt
coreapi==2.3.3
django-autocomplete-light==3.3.2
django-cors-headers==2.5.2
django-debug-toolbar==1.11
django-filter==2.1.0
django-map-widgets==0.2.2
Django==2.1.7
djangorestframework-camel-case==1.0.3
djangorestframework-gis==0.14
djangorestframework==3.9.2
elasticsearch>=6.0.0,<6.3.1
gunicorn==19.9.0
psycopg2==2.7.7
Pygments==2.3.1
python-memcached==1.59
wagtail==2.4.0
